Conduit.App.Monad

Documentation

newtype AppM a

The app monad. woo.

Constructors

AppM

Fields

runAppM :: ReaderT Env IO a

class Monad m => MonadApp m where

Lifts computations to the app monad level. (MonadApp m) shouldn't be used directly as a constraint; Rather, just use liftApp and see Has.

Methods

liftApp :: AppM a -> m a

runService :: FeatureError e => AppM (Either e a) -> ActionT AppM a

Runs a service in the App monad and converts its potentially failed result to the appropriate error if neccesary.

data MyErr = Aw Text
data MyResult = Yay Text deriving (Generic, ToJSON)

instance FeatureError MyErr where
  handleFeatureError (ResultErr msg) = do 
    status status500
    text msg

endpoint = get "/" $ do
  (result :: MyResult) <- runService (someAppService :: AppM (Either MyErr MyResult))
  json result