Conduit.DB.Core

Documentation

newtype DBPool

The ConnectionPool managing the DB connections.

Constructors

DBPool

Fields

unPool :: ConnectionPool

class Monad m => MonadDB m where

Some monad which can run an Esqueleto SQL query/stmt.

Methods

runDB :: SqlPersistT m a -> m (Either DBError a)

class SqlKey t id | t -> id, id -> t where

An abstraction to allow for easy conversion between Esqueleto entity Keys and Conduit's own ID datatypes. deriveSqlKey can automagically create instances for this class.

newtype TableID = TableID { unID :: Int64 }
<define some persist table Table>
$(deriveSqlKey ''Table ''TableID)

Methods

sqlKey2ID :: Key t -> id

id2sqlKey :: id -> Key t

deriveSqlKey :: Name -> Name -> Q [Dec]

just tried this for fun, very quickly realized I am nowhere near smart enough to be doing something like this. this took me over an hour. help.

see SqlKey for usage

data DBError

Attempts to map relevant SqlErrors to more processable types.

Constructors

SomeDBError Text
Catch-all for irrelevant/unkown SqlErrors

UniquenessError Text
Unique constraint violation err, holds name of violated column

AuthorizationError Text
See authorizationSqlError

NotFoundError
See resourceNotFoundSqlError

mapDBResult :: FeatureError e => (a -> b) -> Either DBError a -> Either e b

Maps both sides of a potentially errored SQL query.

newtype Name = Name Text

mkNames :: [Value Text] -> [Name]
mkNames = map (\(Value name) -> Name name)

result :: (Monad m, MonadDB m, MonadUnliftIO m) => m (Either ??? [Name])
result = mapDBResult mkNames <$> runDBError do
  select $ do
    u <- from table @User
    pure u.name

mapMaybeDBResult :: FeatureError e => e -> (a -> b) -> Either DBError (Maybe a) -> Either e b

Maps both sides of a potentially errored, potentially not-found SQL query.

mapDBError :: FeatureError e => Either DBError a -> Either e a

Maps the error of a potentially errored SQL query/stmt.

expectDBNonZero :: (FeatureError e, Num cnt, Ord cnt) => e -> Either DBError cnt -> Either e ()

Maps the error of a potentially errored SQL query/stmt & Assurances that the result is > 0. Intended for use with something like Esqueleto's insertCount or deleteCount.

authorizationSqlError :: Show e => e -> SqlError

A custom SqlError w/ code 45401

resourceNotFoundSqlError :: SqlError

A custom SqlError w/ code 45404

catchSqlError :: MonadUnliftIO m => m a -> m (Either DBError a)

(Internal) Catches & maps any SqlErrors to DBErrors

mapSqlError :: SqlError -> DBError

extractUniquenessViolation :: SqlError -> Text

(Internal) Extracts the name of the column whose uniqueness was violated

extractKeyField :: String -> Maybe String

defaultSqlErr :: SqlError

(Internal) SqlError with irrelevant/unused fields pre-filled out

getConName :: Name -> Q Name

(Internal) Gets the constructor name of some newtype