Conduit.DB.Utils

Documentation

dbTimeNow :: UTCTime

Temporarily sets the time to some irrelevant number for typing reasons. Used for rows where the DB sets the timestamp using some trigger.

suchThat :: SqlQuery a -> (a -> SqlExpr (Value Bool)) -> SqlQuery a infixl 5

Shorthand for DB queries that select from a table just applying some filter

select $ do 
  e <- from table `suchThat` \e -> e.id ==. ???
  ...