Conduit.Errors

Documentation

class FeatureError e where

Methods

handleFeatureError :: MonadIO m => e -> ActionT m ()

Converts a feature error into some scotty response

handleDBError :: DBError -> e

Converts a DBError into some feature error

handleFeatureErrors :: (MonadUnliftIO m, FeatureError e) => e -> ActionT m a

Converts a feature error to its appropriate error response. See runService.

class FeatureErrorMapper e1 e2 where

Provides an easy way for errors to translate between features to facilitate cross-feature code sharing

Methods

mapFeatureError :: e1 -> e2