Conduit.Features.Articles.DB

Documentation

migrateArticleTables :: Migration

createArticleFunctions :: MonadIO m => SqlPersistT m ()

assumingUserIsOwner :: forall table id m a e. (OwnableEntity table, SqlKey table id, MonadIO m, Show e) => e -> UserID -> id -> SqlPersistT m a -> SqlPersistT m a

mkManyArticles :: [(Entity Article, Entity User, Value Bool, Value Bool, Value Int)] -> ManyArticles

mkOneArticle :: (Entity Article, Entity User, Value Bool, Value Bool, Value Int) -> OneArticle

data Article

Constructors

Article

Fields

articleAuthor :: !UserId

articleSlug :: !Text

articleTitle :: !Text

articleDesc :: !Text

articleBody :: !Text

articleTags :: ![Text]

articleCreated :: !UTCTime

articleUpdated :: !UTCTime

data Favorite

Constructors

Favorite

Fields

favoriteUser :: !UserId

favoriteArticle :: !ArticleId

data Comment

Constructors

Comment

Fields

commentAuthor :: !UserId

commentArticle :: !ArticleId

commentBody :: !Text

commentCreated :: !UTCTime

commentUpdated :: !UTCTime

Orphan instances

SqlKey Article ArticleID

Instance details

Methods

sqlKey2ID :: Key Article -> ArticleID

id2sqlKey :: ArticleID -> Key Article

SqlKey Comment CommentID

Instance details

Methods

sqlKey2ID :: Key Comment -> CommentID

id2sqlKey :: CommentID -> Key Comment