Conduit.Identity.Auth

Documentation

data AuthedUser

The form of an authenticated user passed into any endpoint using withAuth/maybeWithAuth.

Constructors

AuthedUser

Fields

authedToken :: !Text

authedUserID :: !UserID

withAuth :: (MonadIO m, MonadReader c m, Has JWTInfo c m) => (AuthedUser -> ActionT m ()) -> ActionT m ()

An endpoint which requires user authentication.

endpoint = get "/" $ withAuth \(user :: AuthedUser) -> do
  ...

maybeWithAuth :: (MonadIO m, Has JWTInfo c m) => (Maybe AuthedUser -> ActionT m ()) -> ActionT m ()

An endpoint which requests, but does require, user authentication.

endpoint = get "/" $ maybeWithAuth \(user :: Maybe AuthedUser) -> do
  ...

class Monad m => AuthTokenGen m where

Some monad which can generate a JWT

Methods

mkAuthToken :: UserID -> m Text

makeAuthTokenPure :: JWTInfo -> POSIXTime -> UserID -> Text

tryMakeAuthedUser :: ToText a => JWTInfo -> POSIXTime -> a -> Maybe AuthedUser

tryGetSubjectFromJWT :: JWTInfo -> POSIXTime -> Text -> Maybe UserID

tryGetClaims :: JWTInfo -> Text -> Maybe JWTClaimsSet

extractToken :: Text -> Maybe Text

authErrRes :: Map Text Text